+++ src/main/java/piglin/swapswap/domain/member/constant/MemberRoleEnum.java
@@ -19,8 +19,8 @@ public enum MemberRoleEnum {
 
     public static class Authority {
 
-        public static final String USER = "ROLE_USER";
-        public static final String ADMIN = "ROLE_ADMIN";
+        public static final String USER = "USER";
+        public static final String ADMIN = "ADMIN";
 
     }
 }
