+++ src/main/resources/templates/coupon/createCouponForm.html
@@ -17,7 +17,7 @@
       event.target.classList.add('active')
     }
   </script>
-  <form th:action="@{/coupons}" th:object="${couponCreateRequestDto}" method="post"
+  <form th:action="@{/admin/coupons}" th:object="${couponCreateRequestDto}" method="post"
         enctype="application/x-www-form-urlencoded">
     <div class="write-bar-container">
       <div class="write-bar">
